+++ R/SPIC.R
@@ -1,5 +1,9 @@
 #' @param fullSeq Logical specifying whether to list all taxa (`TRUE`), or
 #' only those that improve information content when all are dropped (`FALSE`).
+#' @param p Proportion of trees that must contain a split before it is included
+#' in the consensus under consideration.  0.5, the default, corresponds to a
+#' majority rule tree; 1.0 will maximize the information content of the
+#' strict consensus.
 #' @inheritParams TipInstability
 #' @describeIn RogueTaxa Shortcut to 'fast' heuristic, with option to return
 #' evaluation of all taxa using `fullSeq = TRUE`.
@@ -7,12 +11,13 @@
 #'
 #' QuickRogue(trees, fullSeq = TRUE)
 #' @importFrom cli cli_progress_bar cli_progress_update cli_progress_done
-#' @importFrom TreeDist ConsensusInfo
+#' @importFrom TreeDist ConsensusInfo SplitwiseInfo ClusteringInfo
 #' @importFrom fastmatch %fin%
-#' @importFrom TreeTools NTip SplitFrequency
+#' @importFrom TreeTools NTip SplitFrequency PectinateTree DropTipPhylo
 #' @export
 QuickRogue <- function (trees,
                         info = 'phylogenetic',
+                        p = 0.5,
                         log = TRUE, average = 'median', deviation = 'mad',
                         neverDrop, fullSeq = FALSE) {
   if (!is.na(pmatch(tolower(info), 'spic'))) {
@@ -26,7 +31,7 @@ QuickRogue <- function (trees,
                         taxNum = NA_character_,
                         taxon = NA_character_,
                         rawImprovement = NA_real_,
-                        IC = ConsensusInfo(c(trees), info = info),
+                        IC = ConsensusInfo(c(trees), info = info, p = p),
                         stringsAsFactors = FALSE))
     }
     if (!is.list(trees)) {
@@ -44,12 +49,30 @@ QuickRogue <- function (trees,
   nKeep <- length(neverDrop)
   candidates <- character(nTip - 2L - nKeep)
   score <- double(nTip - 2)
-  score[1] <- ConsensusInfo(trees, info = info, check.tips = FALSE)
+  score[1] <- ConsensusInfo(trees, info = info, p = p, check.tips = FALSE)
   nDrops <- nTip - 3L - nKeep
-  cli_progress_bar("Dropping leaves", total = nDrops * (nDrops + 1L) / 2)
-  for (i in 1 + seq_len(nDrops)) {
+  TotalInfo <- switch(pmatch(info, c('phylogenetic', 'clustering')),
+                      SplitwiseInfo, ClusteringInfo)
+
+  cli_progress_bar("Drop leaf", total = nDrops * (nDrops + 1L) / 2,
+                   .auto_close = FALSE)
+  for (i in 1L + seq_len(nDrops)) {
+    bestPossibleNext <- TotalInfo(PectinateTree(nTip - i))
+    bestYet <- max(score, na.rm = TRUE)
+    if (bestPossibleNext < bestYet) {
+      # message("Broken out: can't attain ", signif(score[i]), " bits with ",
+      #         nTip - i, " leaves.")
+      break
+    }
+    bitStat <- if (bestPossibleNext < 1000) {
+      paste0(round(bestYet), '/', round(bestPossibleNext), ' bits')
+    } else if (bestPossibleNext < 1000000) {
+      paste0(round(bestYet / 1000), '/', round(bestPossibleNext / 1000), ' kb')
+    } else {
+      paste0(round(bestYet / 1e6), '/', round(bestPossibleNext / 1e6), ' Mb')
+    }
     cli_progress_update(nDrops - (i - 1),
-                        status = paste0("Leaf ", i - 1, "/", nDrops))
+                        status = paste0("Leaf ", i - 1, '; ', bitStat))
     tipScores <- TipInstability(tr, log = log, average = average,
                                 deviation = deviation,
                                 checkTips = FALSE)
@@ -58,8 +81,8 @@ QuickRogue <- function (trees,
     if (length(candidate)) {
       candidates[i] <- names(candidate)
     }
-    tr <- lapply(tr, DropTip, candidate, preorder = FALSE)
-    score[i] <- ConsensusInfo(tr, info = info, check.tips = FALSE)
+    tr <- lapply(tr, DropTipPhylo, candidate, preorder = FALSE, check = FALSE)
+    score[i] <- ConsensusInfo(tr, info = info, p = p, check.tips = FALSE)
   }
   cli_progress_done()
 
@@ -68,11 +91,12 @@ QuickRogue <- function (trees,
   pointer <- bestPos - 1L
   needsRecalc <- logical(length(candidates))
 
-  cli_progress_bar("Restoring leaves", total = bestPos - 2L)
+  cli_progress_bar("Restore leaf", total = bestPos - 2L)
   while (pointer > 1L) {
-    tryScore <- ConsensusInfo(lapply(trees, DropTip,
-                         candidates[seq_len(bestPos)[-c(1, pointer)]]),
-                  info = info, check.tips = FALSE)
+    tryScore <- ConsensusInfo(
+      lapply(trees, DropTipPhylo, candidates[seq_len(bestPos)[-c(1, pointer)]],
+             preorder = FALSE, check = FALSE),
+      info = info, p = p, check.tips = FALSE)
     if (tryScore > bestScore) {
       candidates[1:bestPos] <- candidates[c((1:bestPos)[-pointer], pointer)]
       bestScore <- tryScore
@@ -84,8 +108,10 @@ QuickRogue <- function (trees,
     pointer <- pointer - 1L
   }
   for (i in which(needsRecalc)) {
-    score[i] <- ConsensusInfo(lapply(trees, DropTip, candidates[seq_len(i)[-1]]),
-                              info = info, check.tips = FALSE)
+    score[i] <- ConsensusInfo(lapply(trees, DropTipPhylo,
+                                     preorder = FALSE, check = FALSE,
+                                     candidates[seq_len(i)[-1]]),
+                              info = info, p = p, check.tips = FALSE)
   }
   cli_progress_done()
 
@@ -99,7 +125,7 @@ QuickRogue <- function (trees,
 
   # Return:
   data.frame(num = seq_along(score) - 1L,
-             taxNum = c(NA_character_, fmatch(dropped, trees[[1]]$tip.label)),
+             taxNum = c(NA_character_, match(dropped, trees[[1]]$tip.label)),
              taxon = c(NA_character_, dropped),
              rawImprovement = c(NA_real_, score[-1] - score[-length(score)]),
              IC = score,
@@ -110,17 +136,17 @@ QuickRogue <- function (trees,
 #' cli_alert_success
 #' @importFrom fastmatch fmatch
 #' @importFrom TreeDist ConsensusInfo
-#' @importFrom TreeTools DropTip SplitFrequency Preorder RenumberTips
+#' @importFrom TreeTools DropTipPhylo SplitFrequency Preorder RenumberTips
 #' @importFrom utils combn
 Roguehalla <- function (trees, dropsetSize = 1, info = 'phylogenetic',
-                        neverDrop) {
+                        p = 0.5, neverDrop) {
   if (!inherits(trees, 'multiPhylo')) {
     if (inherits(trees, 'phylo')) {
       return(data.frame(num = 0,
                         taxNum = NA_character_,
                         taxon = NA_character_,
                         rawImprovement = NA_real_,
-                        IC = ConsensusInfo(c(trees), info = info),
+                        IC = ConsensusInfo(c(trees), info = info, p = p),
                         stringsAsFactors = FALSE))
     }
     if (!is.list(trees)) {
@@ -133,11 +159,10 @@ Roguehalla <- function (trees, dropsetSize = 1, info = 'phylogenetic',
   startTrees <- trees
   labels <- startTrees[[1]]$tip.label
   nTree <- length(trees)
-  majority <- 0.5 + sqrt(.Machine$double.eps)
 
   startTip <- NTip(trees[[1]])
   neverDrop <- .NeverDrop(neverDrop, trees[[1]]$tip.label)
-  best <- ConsensusInfo(trees, info = info, check.tips = FALSE)
+  best <- ConsensusInfo(trees, info = info, p = p, check.tips = FALSE)
 
   .Drop <- function (n) {
     cli_progress_bar(paste0("Dropset size ", n))
@@ -151,8 +176,9 @@ Roguehalla <- function (trees, dropsetSize = 1, info = 'phylogenetic',
       cli_progress_update(1, .envir = parent.frame(2), status = paste0(
         "Drop ", startTip - NTip(trees[[1]]), " leaves = ",
         signif(best), " bits."))
-      dropForest <- lapply(trees, DropTip, drop, preorder = FALSE)
-      ConsensusInfo(dropForest, info = info, check.tips = FALSE)
+      dropForest <- lapply(trees, DropTipPhylo, drop,
+                           check = FALSE, preorder = FALSE)
+      ConsensusInfo(dropForest, info = info, p = p, check.tips = FALSE)
     })
     cli_progress_done()
     if (max(candidates) > best) {
@@ -177,7 +203,8 @@ Roguehalla <- function (trees, dropsetSize = 1, info = 'phylogenetic',
         dropSeq <- c(dropSeq, paste0(thisDrop, collapse = ','))
         taxSeq <- c(taxSeq, paste0(fmatch(thisDrop, labels), collapse = ','))
         dropInf <- c(dropInf, best)
-        trees <- lapply(trees, DropTip, thisDrop, preorder = FALSE)
+        trees <- lapply(trees, DropTipPhylo, thisDrop, preorder = FALSE,
+                        check = FALSE)
         break
       }
     }