+++ tests/testthat/test-stability.R
@@ -1,18 +1,22 @@
-library('TreeTools', quietly = TRUE)
+library('TreeTools', quietly = TRUE, warn.conflicts = FALSE)
 
-test_that("Errors handled", {
-  expect_error(TipInstability(BalancedTree(4)))
+test_that("TipInstability() error handling", {
   expect_error(TipInstability(as.phylo(0:3, 6), dev = 'error'))
   expect_error(TipInstability(as.phylo(0:3, 6), ave = 'error'))
 })
 
-test_that("Cophenetic() works", {
+test_that("TipInstability() null output", {
+  expect_equal(TipInstability(BalancedTree(4)),
+               setNames(rep(0, 4), paste0('t', 1:4)))
+})
+
+test_that("GraphGeodesic() works", {
   Test <- function (tr) {
     tr <- Preorder(tr)
     tr$edge.length <- rep(1, nrow(tr$edge))
     tips <- seq_along(tr$tip.label)
-    expect_equal(unname(ape::dist.nodes(tr)[tips, tips]), Cophenetic(tr))
-    expect_equal(log(Cophenetic(tr)), Cophenetic(tr, log = TRUE))
+    expect_equal(unname(ape::dist.nodes(tr)[tips, tips]), GraphGeodesic(tr))
+    expect_equal(log(GraphGeodesic(tr)), GraphGeodesic(tr, log = TRUE))
   }
   Test(BalancedTree(4))
   Test(BalancedTree(6))