+++ NEWS.md
@@ -1,7 +1,21 @@
-# RogueNaRok v2.0.0
+# Rogue v2.1.0
 
-- Information theoretic rogue detection (per Smith, forthcoming).
+- Early termination of `QuickRogue()` when no further improvement possible.
 
-# RogueNaRok v1.0.0
+- `Cophenetic()` renamed to the more accurate `GraphGeodesic()`.
+
+- Calculate information content of consensus trees with p > 0.5 (#15).
+
+- Improve support for `multiPhylo` objects.
+
+- New vignette detailing rogue detection with Bayesian tree samples.
+
+
+# Rogue v2.0.0
+
+- Information theoretic rogue detection (per Smith, 2022).
+
+
+# Rogue v1.0.0
 
  - R interface to RogueNaRok.